requirements.txt
anndata==0.10.8
click==8.1.7
lmdb==1.4.1
numpy==1.22.4
pandas==1.5.3
PyYAML==6.0.1
PyYAML==6.0.1
scanpy==1.10.2
scikit_learn==1.0.2
scipy==1.14.0
tabulate==0.9.0
torch==2.1.2
tqdm==4.66.1
